# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/memorials": {
      "post": {
        "operationId": "MemorialController_createMemorial",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemorialRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMemorialResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "Memorial"
        ]
      },
      "get": {
        "operationId": "MemorialController_getAllMemorials",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadMemorialResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "Memorial"
        ]
      }
    },
    "/memorials/{id}": {
      "patch": {
        "operationId": "MemorialController_updateMemorial",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemorialRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMemorialResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "Memorial"
        ]
      },
      "delete": {
        "operationId": "MemorialController_deleteMemorial",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadMemorialResponseDto"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "Memorial"
        ]
      }
    },
    "/memorials/readById/{id}": {
      "get": {
        "operationId": "MemorialController_getMemorialById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadMemorialResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "Memorial"
        ]
      }
    },
    "/memorials/creator/{creatorId}/{memorialId}": {
      "get": {
        "operationId": "MemorialController_getMemorialsByCreatorIdAndMode",
        "parameters": [
          {
            "name": "creatorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memorialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "Memorial"
        ]
      }
    },
...
    "/user-media/memorial/{memorialId}": {
      "get": {
        "operationId": "UserMediaController_getMediaByMemorialId",
        "parameters": [
          {
            "name": "memorialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserMediaResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserMedia"
        ]
      }
    },
...
    "/user-tributes/memorial/{memorialId}": {
      "get": {
        "operationId": "UserTributeController_getTributesByMemorialId",
        "parameters": [
          {
            "name": "memorialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserTributeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserTribute"
        ]
      }
    },
...
    "/memorial-qr-codes": {
      "post": {
        "operationId": "MemorialQRCodeController_createQRCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemorialQRCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMemorialQRCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "MemorialQRCode"
        ]
      }
    },
    "/memorial-qr-codes/readById/{id}": {
      "get": {
        "operationId": "MemorialQRCodeController_getQRCodeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadMemorialQRCodeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "MemorialQRCode"
        ]
      }
    },
    "/memorial-qr-codes/memorial/{memorialId}": {
      "get": {
        "operationId": "MemorialQRCodeController_getQRCodeByMemorialId",
        "parameters": [
          {
            "name": "memorialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadMemorialQRCodeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "MemorialQRCode"
        ]
      }
    },
    "/memorial-qr-codes/{id}": {
      "patch": {
        "operationId": "MemorialQRCodeController_updateQRCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemorialQRCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMemorialQRCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "MemorialQRCode"
        ]
      },
      "delete": {
        "operationId": "MemorialQRCodeController_deleteQRCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadMemorialQRCodeResponseDto"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "MemorialQRCode"
        ]
      }
    },
...
  },
  "info": {
    "title": "rememeber-me",
    "description": "API documentation for rememeber-me.env",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateMemorialRequestDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Celebrating the life of John Doe"
          },
          "favSaying": {
            "type": "string",
            "example": "“To live in hearts we leave behind is not to die.”"
          },
          "creatorId": {
            "type": "string",
            "example": "uuid-of-user",
            "description": "Will be automatically set from authenticated user if not provided"
          },
          "landingModeId": {
            "type": "string",
            "example": "uuid-of-landing-mode"
          },
          "templateId": {
            "type": "string",
            "example": "uuid-of-template"
          },
          "personName": {
            "type": "string",
            "example": "John Doe"
          },
          "personGender": {
            "type": "string",
            "example": "male",
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ],
            "description": "Gender of the deceased"
          },
          "personBirthDate": {
            "format": "date-time",
            "type": "string",
            "example": "1950-05-10",
            "description": "Birthdate (ISO format)"
          },
          "personDeathDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-07-15",
            "description": "Death date (ISO format)"
          },
          "profilePictureId": {
            "type": "string",
            "example": "uuid-of-profile-picture"
          },
          "personProfilePicture": {
            "type": "string",
            "example": "https://cdn.example.com/photos/john.jpg"
          },
          "favQuote": {
            "type": "string",
            "example": "“To live in hearts we leave behind is not to die.”"
          },
          "featuredPhotoId": {
            "type": "string",
            "example": "uuid-of-featured-photo"
          },
          "featuredPhotoURL": {
            "type": "string",
            "example": "https://cdn.example.com/photos/featured.jpg"
          },
          "lifeStoryText": {
            "type": "string",
            "example": "John was a wonderful father and friend..."
          },
          "lifeStoryImageId": {
            "type": "string",
            "example": "uuid-of-life-story-image"
          },
          "lifeStoryImageURL": {
            "type": "string",
            "example": "https://cdn.example.com/photos/life-story.jpg"
          },
          "publishStatus": {
            "type": "string",
            "example": "draft",
            "enum": [
              "draft",
              "standard",
              "private",
              "archived"
            ],
            "default": "draft"
          },
          "pageURL": {
            "type": "string",
            "example": "https://rememberme.com/memorial/john-doe"
          },
          "qrCodeData": {
            "type": "string",
            "example": "https://rememberme.com/memorial/john-doe",
            "description": "QR code data (usually the page URL). QR code will be created automatically if provided."
          },
          "eventStart": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T18:00:00Z"
          },
          "eventDuration": {
            "type": "string",
            "example": "48h",
            "description": "Duration in hours/days"
          },
          "autoRevertToFullMode": {
            "type": "boolean",
            "example": true
          },
          "userMedia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateUserMediaRequestDto"
            }
          },
          "userTributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateUserTributeRequestDto"
            }
          },
          "favoriteSayings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMemorialSayingRequestDto"
            }
          }
        },
        "required": [
          "landingModeId",
          "templateId"
        ]
      },
      "CreateMemorialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "UpdateMemorialRequestDto": {
        "type": "object",
        "properties": {
          "favSaying": {
            "type": "string",
            "example": "\"To live in hearts we leave behind is not to die.\""
          },
          "landingModeId": {
            "type": "string",
            "example": "uuid-of-landing-mode"
          },
          "templateId": {
            "type": "string",
            "example": "uuid-of-template"
          },
          "personName": {
            "type": "string",
            "example": "John Doe"
          },
          "personBirthDate": {
            "format": "date-time",
            "type": "string",
            "example": "1950-05-10",
            "description": "Birthdate (ISO format)"
          },
          "personDeathDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-07-15",
            "description": "Death date (ISO format)"
          },
          "profilePictureId": {
            "type": "string",
            "example": "uuid-of-profile-picture"
          },
          "personProfilePicture": {
            "type": "string",
            "example": "https://cdn.example.com/photos/john.jpg"
          },
          "favQuote": {
            "type": "string",
            "example": "\"To live in hearts we leave behind is not to die.\""
          },
          "eventStart": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T18:00:00Z"
          },
          "eventDuration": {
            "type": "string",
            "example": "48h",
            "description": "Duration in hours/days"
          },
          "autoRevertToFullMode": {
            "type": "boolean",
            "example": true
          },
          "userMedia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateUserMediaRequestDto"
            }
          },
          "userTributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateUserTributeRequestDto"
            }
          }
        }
      },
      "UpdateMemorialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "ReadUserMediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-media"
          },
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "fileURL": {
            "type": "string",
            "example": "https://cdn.example.com/uploads/media-file.png"
          },
          "fileId": {
            "type": "string",
            "example": "uuid-of-file"
          },
          "type": {
            "type": "string",
            "example": "IMAGE",
            "description": "Media type (e.g., IMAGE, VIDEO, AUDIO)"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "photoCaption": {
            "type": "string",
            "example": "Vacation Memories"
          },
          "photoDescription": {
            "type": "string",
            "example": "A beautiful sunset at the beach"
          },
          "videoTitle": {
            "type": "string",
            "example": "Grandma’s 80th Birthday Celebration"
          },
          "isMainVideo": {
            "type": "boolean",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "mimeType",
          "fileURL",
          "type",
          "memorialId",
          "userId",
          "isMainVideo",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadMemorialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "description": {
            "type": "string",
            "example": "Celebrating the life of John Doe"
          },
          "favSaying": {
            "type": "string",
            "example": "John was a wonderful father and friend."
          },
          "creatorId": {
            "type": "string",
            "example": "uuid-of-creator"
          },
          "landingModeId": {
            "type": "string",
            "example": "uuid-of-landing-mode"
          },
          "landingMode": {
            "$ref": "#/components/schemas/ReadLandingModeResponseDto"
          },
          "templateId": {
            "type": "string",
            "example": "uuid-of-template"
          },
          "personName": {
            "type": "string",
            "example": "John Doe"
          },
          "personGender": {
            "type": "string",
            "example": "male",
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ]
          },
          "personBirthDate": {
            "format": "date-time",
            "type": "string",
            "example": "1950-01-01T00:00:00Z"
          },
          "personDeathDate": {
            "format": "date-time",
            "type": "string",
            "example": "2020-05-15T00:00:00Z"
          },
          "personProfilePicture": {
            "type": "string",
            "example": "https://cdn.example.com/photos/john.png"
          },
          "profilePictureId": {
            "type": "string",
            "example": "uuid-of-profile-picture"
          },
          "favQuote": {
            "type": "string",
            "example": "“Live life to the fullest.”"
          },
          "featuredPhotoId": {
            "type": "string",
            "example": "uuid-of-featured-photo"
          },
          "featuredPhotoURL": {
            "type": "string",
            "example": "https://cdn.example.com/photos/featured.jpg"
          },
          "lifeStoryText": {
            "type": "string",
            "example": "John was a wonderful father and friend..."
          },
          "lifeStoryImageId": {
            "type": "string",
            "example": "uuid-of-life-story-image"
          },
          "lifeStoryImageURL": {
            "type": "string",
            "example": "https://cdn.example.com/photos/life-story.jpg"
          },
          "publishStatus": {
            "type": "string",
            "example": "draft",
            "enum": [
              "draft",
              "standard",
              "private",
              "archived"
            ]
          },
          "eventStart": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T10:00:00Z"
          },
          "eventDuration": {
            "type": "string",
            "example": "48h"
          },
          "autoRevertToFullMode": {
            "type": "boolean",
            "example": true
          },
          "slug": {
            "type": "string",
            "example": "john-doe-memorial"
          },
          "pageURL": {
            "type": "string",
            "example": "https://rememberme.com/memorial/john-doe"
          },
          "landingModeType": {
            "type": "string",
            "example": "full-mode",
            "enum": [
              "full-mode",
              "event-mode",
              "video-only-mode"
            ],
            "description": "Memorial mode type from landing mode"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-29T09:21:45Z"
          },
          "userMedia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadUserMediaResponseDto"
            }
          },
          "userTributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadUserTributeResponseDto"
            }
          },
          "favoriteSayings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadMemorialSayingResponseDto"
            }
          },
          "qrCode": {
            "$ref": "#/components/schemas/ReadMemorialQRCodeResponseDto"
          }
        },
        "required": [
          "id",
          "creatorId",
          "landingModeId",
          "templateId",
          "autoRevertToFullMode",
          "landingModeType",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateMemorialQRCodeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-qr-code"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "UpdateMemorialQRCodeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-qr-code"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}